requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
Jinja2
starlette
pydantic
requests
gunicorn
python-multipart
python-dotenv